refactor(cli): extract work board/roster/manager renderers into work_commands_render.rs (card c0bd865c phase 1)
## Why

Joel pushback 2026-05-28: 'pretend this is a project someone would actually use / not a SHIM based codebase'. work_commands.rs at 1880 LOC, 45 free functions, 0 structs/traits — embedders (continuum, hermes, openclaw, codex) can't read the API surface in one pass.

## What this PR lands

Phase 1 of the 5-way split. The 5 print_* functions (terminal output formatters) move to a sibling module `work_commands_render.rs`:

  - print_board
  - print_work_queue_availability
  - print_work_roster
  - print_work_manager
  - print_manager_recommendation

These were the lowest-risk extraction:
  - Pure terminal formatting, no substrate calls, no state mutation
  - Single set of inputs (passed in by handlers in work_commands.rs)
  - No cross-function coupling beyond the small format_* helpers

Module declared in `main.rs`. Cross-module visibility: `format_lease`, `format_peer`, `short_id`, `now_ms`, `BoardFilter::matches` made `pub(crate)` so render can call them. Handlers in work_commands.rs now route through `crate::work_commands_render::print_*`.

## File sizes

  work_commands.rs:        1880 → 1642 LOC (-238)
  work_commands_render.rs:    0 →  258 LOC (new)

## Test coverage preserved

All 20 existing work_commands tests still pass. No tests moved — they exercise handlers + helpers, both of which stay in work_commands.rs.

## What this is NOT yet

The card title is the FULL 5-way split: lifecycle / git / gh / board / review. This PR ships phase 1 (render). Phases 2-5 land as follow-up cards so each is a focused review surface:

  - PHASE 2 — lifecycle: run_create/seed/claim/release/heartbeat/update/state/close + their state guards
  - PHASE 3 — git: spawn_claim_worktree + git_* helpers (will be ShellGitClient consumers once dec35ec7 lands)
  - PHASE 4 — gh: open_pr_and_link + gh_default_branch + pr_create_base_branch (likewise ShellGhClient consumers)
  - PHASE 5 — review: run_review + auto_spawn_review_card + format_review_title

Sequencing rationale: render was the lowest-coupling group. Each subsequent phase reduces work_commands.rs by a similar margin. After phase 5, work_commands.rs should be ~400 LOC of pure handlers + shared types.

## Anchor

Other tab's c0bd865c framing + Joel directive 2026-05-28 ('pile of shit / make embeddable'). Pair-mate with dec35ec7 (typed clients via airc-lib): when those land the git_* and gh_* modules in subsequent phases migrate to trait calls instead of raw subprocess.
