+++ crates/airc-cli/src/work_commands.rs
@@ -15,8 +15,7 @@ use uuid::Uuid;
 use airc_lib::{
     AgentAvailabilityState, CardState, ChangeWorkCardState, ClaimId, ClaimWorkCard, CreateWorkCard,
     LaneId, Priority, ReleaseWorkClaim, RepoId, UpdateWorkCard, WorkBacklogSeedCandidate,
-    WorkBacklogSeedOutcome, WorkBoardProjection, WorkCardId, WorkManagerRecommendation,
-    WorkManagerRecommendationKind, WorkManagerStatus, WorkQueueStatus, WorkRosterStatus,
+    WorkBacklogSeedOutcome, WorkCardId,
 };
 
 use crate::lease;
@@ -612,10 +611,23 @@ async fn open_pr_and_link(
     // Card a4fe899f: `gh` does NOT accept `-C` (that's git's flag).
     // The cwd has to be set via `Command::current_dir(...)` so gh's
     // own repo-resolution (which scans `git remote get-url origin`
-    // from cwd) picks the worktree's branch. Without this, gh
-    // silently ran against whatever the shell's cwd was at invoke
-    // time — usually the wrong repo — and the whole Review-state →
-    // PR-link pipeline failed (best-effort warning was swallowed).
+    // from cwd) picks the worktree's branch.
+    //
+    // Card 28f1440c: `--base` is set explicitly to the workflow's
+    // integration branch (`rust-rewrite`) rather than letting gh
+    // fall back to the repo's GitHub default (`main` here). Main is
+    // the legacy snapshot; every PR landing there bypasses the
+    // substrate work. Hardcoded for the MVP; a follow-up card
+    // surfaces it as configurable for consumers whose integration
+    // branch differs.
+    //
+    // Card 13131f1c: `--title` + `--body` are passed explicitly from
+    // the HEAD commit instead of relying on `--fill`. `--fill`'s
+    // heuristic sometimes substitutes a slugified branch name as the
+    // title (observed on #1032 and #1033 — `docs(agents):`/`fix(ci):`
+    // prefixes stripped, breaking the canary-gate bypass). Reading
+    // the commit subject directly is deterministic.
+    let base_branch = pr_create_base_branch();
     let subject = git_show_format(&worktree_str, "%s")?;
     let body = git_show_format(&worktree_str, "%b")?;
     let create_out = std::process::Command::new("gh")
@@ -627,6 +639,8 @@ async fn open_pr_and_link(
             subject.trim(),
             "--body",
             body.trim(),
+            "--base",
+            base_branch,
         ])
         .output()?;
     if !create_out.status.success() {
@@ -641,15 +655,18 @@ async fn open_pr_and_link(
     let pr_number = extract_pr_number(pr_url)
         .ok_or_else(|| format!("could not parse PR number from gh output: {pr_url}"))?;
 
-    // Resolve head/base from the worktree's git state.
+    // Resolve head from the worktree's git state. Base is the same
+    // value we passed to `gh pr create` above (card 28f1440c) — must
+    // not drift between the substrate's PullRequestRef and the
+    // actual PR target, or downstream consumers (review-spawn,
+    // close-guard) reason about the wrong branch.
     let head_branch = git_rev_parse_branch(&worktree_str)?;
-    let base_branch = gh_default_branch(&worktree_str).unwrap_or_else(|_| "main".to_string());
 
     let pull_request = PullRequestRef {
         repo: card.repo.clone(),
         number: pr_number,
         head: BranchName::new(head_branch)?,
-        base: BranchName::new(base_branch)?,
+        base: BranchName::new(base_branch.to_string())?,
     };
     airc.link_card_pull_request(airc_lib::LinkCardPullRequest {
         card_id,
@@ -786,6 +803,22 @@ fn git_show_format(worktree: &str, format: &str) -> Result<String, Box<dyn std::
     Ok(String::from_utf8(out.stdout)?)
 }
 
+/// Card 28f1440c — the integration branch every per-card PR opens
+/// against. Hardcoded to the airc substrate's working branch for
+/// MVP; a follow-up surfaces this as configurable (e.g. via
+/// `.airc/work.toml`) for consumers whose integration branch is
+/// different.
+///
+/// MUST NOT change to a runtime read of `gh repo view --json defaultBranchRef` —
+/// that returns the repo's GitHub `default_branch` (`main` on this
+/// repo today), and every PR landing on `main` bypasses
+/// `rust-rewrite`'s substrate work. The whole point of this card
+/// is to refuse that fallback.
+pub(crate) fn pr_create_base_branch() -> &'static str {
+    "rust-rewrite"
+}
+
+#[allow(dead_code)]
 fn gh_default_branch(worktree: &str) -> Result<String, Box<dyn std::error::Error>> {
     // Card a4fe899f: `gh` does NOT accept `-C`; set cwd via
     // `Command::current_dir(...)` so `gh repo view` resolves the
@@ -898,7 +931,7 @@ pub(crate) fn close_transition_allowed_from(state: CardState) -> bool {
 /// We deliberately do NOT shorten card_id in the board output because
 /// callers copy-paste it into `claim` / `state` / `close`; it's the
 /// API key, not a display field.
-fn short_id<T: std::fmt::Display>(id: T) -> String {
+pub(crate) fn short_id<T: std::fmt::Display>(id: T) -> String {
     id.to_string().chars().take(8).collect()
 }
 
@@ -908,7 +941,7 @@ fn short_id<T: std::fmt::Display>(id: T) -> String {
 /// short-uuid fallback. Honest "unknown" rendering keeps the board
 /// readable even when a peer hasn't published an identity card to
 /// this room yet.
-fn format_peer(
+pub(crate) fn format_peer(
     peer: airc_lib::PeerId,
     me: airc_lib::PeerId,
     aliases: &std::collections::HashMap<airc_lib::PeerId, String>,
@@ -948,7 +981,12 @@ impl BoardFilter {
         }
     }
 
-    fn matches(self, card: &airc_work::WorkCard, me: airc_lib::PeerId, now_ms: u64) -> bool {
+    pub(crate) fn matches(
+        self,
+        card: &airc_work::WorkCard,
+        me: airc_lib::PeerId,
+        now_ms: u64,
+    ) -> bool {
         match self {
             Self::All => true,
             Self::Available => {
@@ -973,7 +1011,7 @@ impl BoardFilter {
 /// the flywheel-continuity doctrine), otherwise time remaining as
 /// 'Mm SS s'. Lets at-a-glance scanning of the board surface stale
 /// claims without having to cross-reference 'work roster' + ttl.
-fn format_lease(expires_at_ms: Option<u64>, now_ms: u64) -> String {
+pub(crate) fn format_lease(expires_at_ms: Option<u64>, now_ms: u64) -> String {
     match expires_at_ms {
         None => "-".to_string(),
         Some(expires_at_ms) if expires_at_ms <= now_ms => "<STALE>".to_string(),
@@ -1025,7 +1063,7 @@ pub async fn run_board(
         }
     }
 
-    print_board(&board, me, filter, &aliases);
+    crate::work_commands_render::print_board(&board, me, filter, &aliases);
     Ok(())
 }
 
@@ -1066,7 +1104,7 @@ pub async fn run_next(
         }
     }
 
-    print_work_queue_availability(&status);
+    crate::work_commands_render::print_work_queue_availability(&status);
     if !status.active_claims_for_peer.is_empty() {
         println!();
         println!(
@@ -1100,7 +1138,7 @@ pub async fn run_roster(
             active_within_ms,
         })
         .await?;
-    print_work_roster(&status);
+    crate::work_commands_render::print_work_roster(&status);
     Ok(())
 }
 
@@ -1124,7 +1162,7 @@ pub async fn run_manage(
             active_within_ms,
         })
         .await?;
-    print_work_manager(&status);
+    crate::work_commands_render::print_work_manager(&status);
     Ok(())
 }
 
@@ -1148,249 +1186,7 @@ pub async fn run_availability(
     Ok(())
 }
 
-fn print_board(
-    board: &WorkBoardProjection,
-    me: airc_lib::PeerId,
-    filter: BoardFilter,
-    aliases: &std::collections::HashMap<airc_lib::PeerId, String>,
-) {
-    let snapshot = board.snapshot();
-    if snapshot.cards.is_empty() && snapshot.agent_availability.is_empty() {
-        println!("(no work cards)");
-        return;
-    }
-    let stale_claims = board.stale_claims(now_ms());
-
-    let now = now_ms();
-    let visible: Vec<&airc_work::WorkCard> = snapshot
-        .cards
-        .iter()
-        .filter(|card| filter.matches(card, me, now))
-        .collect();
-    if !visible.is_empty() {
-        if matches!(filter, BoardFilter::All) {
-            println!("work cards: {}", visible.len());
-        } else {
-            println!(
-                "work cards: {} (filter={:?}, hidden={})",
-                visible.len(),
-                filter,
-                snapshot.cards.len() - visible.len(),
-            );
-        }
-    } else if !matches!(filter, BoardFilter::All) {
-        println!("(no cards match filter {:?})", filter);
-    }
-    for card in &visible {
-        let owner = card
-            .owner
-            .map(|peer| format_peer(peer, me, aliases))
-            .unwrap_or_else(|| "-".to_string());
-        let claim = card
-            .claim_id
-            .map(short_id)
-            .unwrap_or_else(|| "-".to_string());
-        let lease = format_lease(card.claim_expires_at_ms, now);
-        println!(
-            "{card_id}  {priority:?}  {state:?}  owner={owner}  claim={claim}  lease={lease}  repo={repo}  title={title}",
-            card_id = card.card_id,
-            priority = card.priority,
-            state = card.state,
-            repo = card.repo,
-            title = card.title,
-        );
-    }
-    if !stale_claims.is_empty() {
-        println!();
-        println!("stale claims: {}", stale_claims.len());
-        for claim in stale_claims {
-            println!(
-                "{card_id}  owner={owner}  claim={claim_id}  expired_at_ms={expired_at_ms}",
-                card_id = claim.card_id,
-                owner = format_peer(claim.owner, me, aliases),
-                claim_id = short_id(claim.claim_id),
-                expired_at_ms = claim.expired_at_ms,
-            );
-        }
-    }
-    if !snapshot.agent_availability.is_empty() {
-        println!();
-        println!("agent availability: {}", snapshot.agent_availability.len());
-        for availability in snapshot.agent_availability {
-            let stale = availability.expires_at_ms <= now_ms();
-            let note = availability.report.note.as_deref().unwrap_or("-");
-            println!(
-                "{repo}  peer={peer}  state={state:?}  stale={stale}  expires_at_ms={expires_at_ms}  note={note}",
-                repo = availability.report.repo,
-                peer = availability.report.peer,
-                state = availability.report.state,
-                expires_at_ms = availability.expires_at_ms,
-            );
-        }
-    }
-}
-
-fn print_work_queue_availability(status: &WorkQueueStatus) {
-    if status.agent_availability.is_empty() {
-        return;
-    }
-
-    let now_ms = now_ms();
-    println!();
-    println!(
-        "agent availability: ready={} busy={} away={} stale={}",
-        status.ready_count(now_ms),
-        status.busy_count(now_ms),
-        status.away_count(now_ms),
-        status.stale_availability_count(now_ms)
-    );
-    for availability in &status.agent_availability {
-        let stale = availability.expires_at_ms <= now_ms;
-        let note = availability.report.note.as_deref().unwrap_or("-");
-        println!(
-            "{repo}  peer={peer}  state={state:?}  stale={stale}  note={note}",
-            repo = availability.report.repo,
-            peer = availability.report.peer,
-            state = availability.report.state,
-        );
-    }
-}
-
-fn print_work_roster(status: &WorkRosterStatus) {
-    let now_ms = now_ms();
-    if status.rows.is_empty() {
-        println!("work roster: 0 agent(s)");
-        println!("claimable work: {}", status.claimable_count);
-        return;
-    }
-
-    println!(
-        "work roster: {} agent(s) live={} ready={} busy={} away={} stale_availability={} claimable={}",
-        status.rows.len(),
-        status.alive_count(),
-        status.ready_count(now_ms),
-        status.busy_count(now_ms),
-        status.away_count(now_ms),
-        status.stale_availability_count(now_ms),
-        status.claimable_count
-    );
-    for row in &status.rows {
-        let live = row
-            .liveness
-            .as_ref()
-            .map(|liveness| {
-                let client = liveness.client_id.as_deref().unwrap_or("-");
-                let build = liveness.build.as_deref().unwrap_or("-");
-                format!(
-                    "live runtime={} client={} scope={} build={} last_seen_ms={}",
-                    liveness.runtime,
-                    client,
-                    liveness.scope.as_deref().unwrap_or("-"),
-                    build,
-                    liveness.last_seen_ms
-                )
-            })
-            .unwrap_or_else(|| "live=false".to_string());
-        let availability = row
-            .availability
-            .as_ref()
-            .map(|availability| {
-                format!(
-                    "availability={:?} repo={} stale={} note={}",
-                    availability.report.state,
-                    availability.report.repo,
-                    availability.expires_at_ms <= now_ms,
-                    availability.report.note.as_deref().unwrap_or("-")
-                )
-            })
-            .unwrap_or_else(|| "availability=unknown".to_string());
-        println!(
-            "peer={peer}  {live}  {availability}  claims={claims}",
-            peer = row.peer,
-            claims = row.active_claims.len(),
-        );
-        for card in &row.active_claims {
-            println!(
-                "  claim {card_id}  {priority:?}  {state:?}  repo={repo}  title={title}",
-                card_id = card.card_id,
-                priority = card.priority,
-                state = card.state,
-                repo = card.repo,
-                title = card.title,
-            );
-        }
-    }
-}
-
-fn print_work_manager(status: &WorkManagerStatus) {
-    let now_ms = now_ms();
-    println!(
-        "work manager: recommendations={} claimable={} agents={} live={} ready={} busy={} away={}",
-        status.recommendations.len(),
-        status.queue.claimable.len(),
-        status.roster.rows.len(),
-        status.roster.alive_count(),
-        status.roster.ready_count(now_ms),
-        status.roster.busy_count(now_ms),
-        status.roster.away_count(now_ms),
-    );
-    if status.recommendations.is_empty() {
-        println!("action: none");
-        return;
-    }
-    for recommendation in &status.recommendations {
-        print_manager_recommendation(recommendation);
-    }
-}
-
-fn print_manager_recommendation(recommendation: &WorkManagerRecommendation) {
-    let action = match recommendation.kind {
-        WorkManagerRecommendationKind::ClaimWork => "claim-work",
-        WorkManagerRecommendationKind::RecoverStaleClaim => "recover-stale-claim",
-        WorkManagerRecommendationKind::PublishAvailability => "publish-availability",
-        WorkManagerRecommendationKind::SeedBacklog => "seed-backlog",
-        WorkManagerRecommendationKind::Wait => "wait",
-    };
-    let card = recommendation
-        .card
-        .as_ref()
-        .map(|card| {
-            format!(
-                " card={} priority={:?} repo={} title={}",
-                card.card_id, card.priority, card.repo, card.title
-            )
-        })
-        .unwrap_or_default();
-    let stale = recommendation
-        .stale_claim
-        .as_ref()
-        .map(|claim| {
-            format!(
-                " stale_claim={} stale_owner={}",
-                claim.claim_id, claim.owner
-            )
-        })
-        .unwrap_or_default();
-    let agent = recommendation
-        .agent
-        .as_ref()
-        .map(|agent| {
-            format!(
-                " agent={} client={} runtime={} scope={}",
-                agent.peer,
-                agent.client_id.as_deref().unwrap_or("-"),
-                agent.runtime.as_deref().unwrap_or("-"),
-                agent.scope.as_deref().unwrap_or("-"),
-            )
-        })
-        .unwrap_or_default();
-    println!(
-        "action={action} reason={reason:?}{card}{stale}{agent}",
-        reason = recommendation.reason
-    );
-}
-
-fn now_ms() -> u64 {
+pub(crate) fn now_ms() -> u64 {
     std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .map(|duration| duration.as_millis().min(u128::from(u64::MAX)) as u64)
@@ -1819,5 +1615,28 @@ mod tests {
         // The id must appear so the agent can copy-paste it into
         // the corrective command.
         assert!(msg.contains(&card_id.to_string()), "carries the card UUID");
+    }
+
+    /// Card 28f1440c — the PR target branch is hardcoded to the
+    /// substrate's working branch (`rust-rewrite`), NOT the repo's
+    /// GitHub default (`main`). Pinning this catches a regression
+    /// where someone "fixes" the constant by routing through
+    /// `gh_default_branch` (which surfaces `main` for this repo and
+    /// would silently bypass the substrate work the doctrine
+    /// requires per AGENTS.md §8). If a config-driven base ever
+    /// ships (the documented follow-up), this test extends —
+    /// it does not get deleted.
+    #[test]
+    fn pr_create_base_branch_targets_rust_rewrite() {
+        assert_eq!(
+            pr_create_base_branch(),
+            "rust-rewrite",
+            "card 28f1440c: PR target must be the substrate working \
+             branch, never the repo's GitHub default ('main' on this \
+             repo today)"
+        );
+        // Specifically: it must NOT be 'main'. The doctrine treats
+        // main as the legacy snapshot.
+        assert_ne!(pr_create_base_branch(), "main");
     }
 }
